+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.3.3 - Unreleased
+[#70](https://github.com/BrynCooke/buildstructor/issues/70)
+Add the ability to override the visibility of generated builder.
 
 [#3](https://github.com/BrynCooke/buildstructor/issues/3)
 Changed the storage of required but not yet supplied values to use MaybeUninit.