+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.3.3 - Unreleased
+## 0.4.0 - 2022-07-31
 
 [#69](https://github.com/BrynCooke/buildstructor/issues/69)
 Add the ability to override the visibility of generated builder.